Collision fix in hi_insert_uint32_t
We stumbled on a problem about collisions when having hash tables where
the key are integers (e.g. hi_insert_uint32_t()).
The problem is libhashish directly saves the pointer of the key given in
hi_insert_uint32_t(). This address is on the stack.
When a collision occurs, the same address is used to check if the key is
already in the hash table thus refuses to add the key which collide
(returning HI_ERR_DUPKEY).

The patch libhashish_collision_fix.patch is here to fix this problem.
This is done by copying the key to insert on the heap and save this
pointer in the hash table. Memory allocated for the keys are freed on
removing the entry or destroying the hash table.

Furthermore, the unit test has been patched to completely fill the hash
table thus triggering collisions.